refactor(orchestrator): extract shared helpers
