+++ services/local-service/internal/orchestrator/orchestrator_helpers.go
@@ -1,0 +1,159 @@
+package orchestrator
+
+import (
+	"encoding/json"
+	"strings"
+	"time"
+
+	contextsvc "github.com/cialloclaw/cialloclaw/services/local-service/internal/context"
+	"github.com/cialloclaw/cialloclaw/services/local-service/internal/runengine"
+	"github.com/cialloclaw/cialloclaw/services/local-service/internal/storage"
+)
+
+func storageTaskRunRecordFromSnapshotJSON(payload string) (storage.TaskRunRecord, error) {
+	var record storage.TaskRunRecord
+	if err := json.Unmarshal([]byte(payload), &record); err != nil {
+		return storage.TaskRunRecord{}, err
+	}
+	return record, nil
+}
+
+func timelineFromStorage(timeline []storage.TaskStepSnapshot) []runengine.TaskStepRecord {
+	if len(timeline) == 0 {
+		return nil
+	}
+	result := make([]runengine.TaskStepRecord, len(timeline))
+	for index, step := range timeline {
+		result[index] = runengine.TaskStepRecord{
+			StepID:        step.StepID,
+			TaskID:        step.TaskID,
+			Name:          step.Name,
+			Status:        step.Status,
+			OrderIndex:    step.OrderIndex,
+			InputSummary:  step.InputSummary,
+			OutputSummary: step.OutputSummary,
+		}
+	}
+	return result
+}
+
+func cloneTimePointer(value *time.Time) *time.Time {
+	if value == nil {
+		return nil
+	}
+	cloned := *value
+	return &cloned
+}
+
+func notepadSnapshot(item map[string]any) contextsvc.TaskContextSnapshot {
+	return contextsvc.TaskContextSnapshot{
+		Source:    "dashboard",
+		InputType: "text",
+		Text:      stringValue(item, "title", ""),
+		PageTitle: "notepad",
+		AppName:   "dashboard",
+	}
+}
+
+func latestOutputPathFromTasks(tasks []runengine.TaskRecord) string {
+	for _, task := range tasks {
+		for _, artifact := range task.Artifacts {
+			if outputPath := stringValue(artifact, "path", ""); outputPath != "" {
+				return outputPath
+			}
+		}
+		if outputPath := pathFromDeliveryResult(task.DeliveryResult); outputPath != "" {
+			return outputPath
+		}
+		if outputPath := stringValue(task.StorageWritePlan, "target_path", ""); outputPath != "" {
+			return outputPath
+		}
+	}
+	return ""
+}
+
+// snapshotFromTask rebuilds the minimum context snapshot needed for resume and
+// other post-creation flows.
+func snapshotFromTask(task runengine.TaskRecord) contextsvc.TaskContextSnapshot {
+	if !isEmptySnapshot(task.Snapshot) {
+		return cloneTaskSnapshot(task.Snapshot)
+	}
+	return contextsvc.TaskContextSnapshot{
+		Trigger:   task.SourceType,
+		InputType: "text",
+		Text:      originalTextFromTaskTitle(task.Title),
+	}
+}
+
+func cloneTaskSnapshot(snapshot contextsvc.TaskContextSnapshot) contextsvc.TaskContextSnapshot {
+	cloned := snapshot
+	if len(snapshot.Files) > 0 {
+		cloned.Files = append([]string(nil), snapshot.Files...)
+	}
+	return cloned
+}
+
+func isEmptySnapshot(snapshot contextsvc.TaskContextSnapshot) bool {
+	return strings.TrimSpace(snapshot.Source) == "" &&
+		strings.TrimSpace(snapshot.Trigger) == "" &&
+		strings.TrimSpace(snapshot.InputType) == "" &&
+		strings.TrimSpace(snapshot.InputMode) == "" &&
+		strings.TrimSpace(snapshot.Text) == "" &&
+		strings.TrimSpace(snapshot.SelectionText) == "" &&
+		strings.TrimSpace(snapshot.ErrorText) == "" &&
+		len(snapshot.Files) == 0 &&
+		strings.TrimSpace(snapshot.PageTitle) == "" &&
+		strings.TrimSpace(snapshot.PageURL) == "" &&
+		strings.TrimSpace(snapshot.AppName) == "" &&
+		strings.TrimSpace(snapshot.BrowserKind) == "" &&
+		strings.TrimSpace(snapshot.ProcessPath) == "" &&
+		snapshot.ProcessID == 0 &&
+		strings.TrimSpace(snapshot.WindowTitle) == "" &&
+		strings.TrimSpace(snapshot.VisibleText) == "" &&
+		strings.TrimSpace(snapshot.ScreenSummary) == "" &&
+		strings.TrimSpace(snapshot.ClipboardText) == "" &&
+		strings.TrimSpace(snapshot.HoverTarget) == "" &&
+		strings.TrimSpace(snapshot.LastAction) == "" &&
+		snapshot.DwellMillis == 0 &&
+		snapshot.CopyCount == 0 &&
+		snapshot.WindowSwitches == 0 &&
+		snapshot.PageSwitches == 0
+}
+
+func originalTextFromTaskTitle(title string) string {
+	trimmed := strings.TrimSpace(title)
+	for _, prefix := range []string{"确认处理方式：", "改写：", "翻译：", "解释错误：", "解释：", "总结文件：", "总结：", "处理："} {
+		if strings.HasPrefix(trimmed, prefix) {
+			return strings.TrimSpace(strings.TrimPrefix(trimmed, prefix))
+		}
+	}
+	return trimmed
+}
+
+func confirmationTitleFromTask(task runengine.TaskRecord) string {
+	subject := strings.TrimSpace(originalTextFromTaskTitle(task.Title))
+	if subject == "" {
+		subject = "当前任务"
+	}
+	return "确认处理方式：" + subject
+}
+
+// mergeSuggestedDeliveryPreference preserves explicit caller preferences and only
+// falls back to the intent layer's suggested delivery when the caller left the
+// preferred delivery unset.
+func mergeSuggestedDeliveryPreference(preferredDelivery, fallbackDelivery, suggestedDelivery string) (string, string) {
+	if strings.TrimSpace(preferredDelivery) == "" && strings.TrimSpace(suggestedDelivery) != "" {
+		preferredDelivery = suggestedDelivery
+	}
+	return preferredDelivery, fallbackDelivery
+}
+
+// buildPendingExecution creates the minimum delivery plan required to resume a
+// task after authorization. The stored plan must be deterministic and task-
+// centric because waiting_auth can outlive the original request and later needs
+// to restart execution without recomputing delivery intent from transport-only
+// inputs.
+func (s *Service) buildPendingExecution(task runengine.TaskRecord, taskIntent map[string]any) map[string]any {
+	plan := s.delivery.BuildApprovalExecutionPlan(task.TaskID, taskIntent)
+	return s.applyResolvedDeliveryToPlan(task, plan, taskIntent)
+}
