+++ services/local-service/internal/orchestrator/workflow_constants.go
@@ -1,0 +1,14 @@
+package orchestrator
+
+import (
+	"time"
+)
+
+const (
+	executionSegmentInitial     = "initial"
+	executionSegmentResume      = "resume"
+	executionSegmentRestart     = "restart"
+	defaultTaskExecutionTimeout = 95 * time.Second
+	subjectPreviewMaxLength     = 24
+	resultPreviewMaxLength      = 120
+)
